+++ apps/web/src/components/Messages/PreviewList.tsx
@@ -2,14 +2,15 @@ import { Trans } from '@lingui/macro';
 import clsx from 'clsx';
 import { useRouter } from 'next/router';
 import type { FC } from 'react';
-import { useEffect } from 'react';
+import { useCallback, useEffect } from 'react';
 import { MESSAGING_PROVIDER } from 'src/constants';
 import { useAppStore } from 'src/store/app';
-import { useMessagePersistStore, useMessageStore } from 'src/store/message';
+import { useMessageStore } from 'src/store/message';
+import { useXmtpMessagePersistStore } from 'src/store/xmtp-message';
 import { Card, GridItemFour } from 'ui';
 
-import PushPreviewList from './PushPreview';
-import XMTPPreviewList from './XMTPPreview';
+import PUSHPreviewList from './Push/PUSHPreview';
+import XMTPPreviewList from './Xmtp/XMTPPreview';
 
 interface PreviewListProps {
   className?: string;
@@ -19,29 +20,27 @@ interface PreviewListProps {
 const PreviewList: FC<PreviewListProps> = ({ className, selectedConversationKey }) => {
   const router = useRouter();
 
-  const { provider: urlProvider = MESSAGING_PROVIDER.PUSH } = router.query;
+  const chatProvider = useMessageStore((state) => state.chatProvider);
+  const setChatProvider = useMessageStore((state) => state.setChatProvider);
 
   const currentProfile = useAppStore((state) => state.currentProfile);
-  const provider = useMessageStore((state) => state.provider);
-  const setProvider = useMessageStore((state) => state.setProvider);
-  const clearMessagesBadge = useMessagePersistStore((state) => state.clearMessagesBadge);
+  const clearMessagesBadge = useXmtpMessagePersistStore((state) => state.clearMessagesBadge);
 
-  const changeProvider = (newProvider: string) => {
-    setProvider(newProvider);
-    const queryParams = { provider: newProvider };
-    if (!router.pathname.includes('[...conversationKey]')) {
-      router.push({
-        pathname: router.pathname,
-        query: queryParams
-      });
-    }
-  };
+  const changeChatProvider = useCallback(
+    (provider: string) => {
+      setChatProvider(provider);
+      const currentPath = router.pathname; // get the current path
+      let newPath = `/messages/${provider}`; // set the new path based on the provider
 
-  useEffect(() => {
-    const defaultProvider =
-      urlProvider === MESSAGING_PROVIDER.XMTP ? MESSAGING_PROVIDER.XMTP : MESSAGING_PROVIDER.PUSH;
-    changeProvider(defaultProvider);
-  }, [urlProvider]);
+      // check if the current path already contains the provider
+      if (currentPath.includes(provider)) {
+        newPath = currentPath; // set the new path to the current path if it already contains the provider
+      }
+
+      router.push(newPath); // update the URL with the new path
+    },
+    [router, setChatProvider]
+  );
 
   useEffect(() => {
     if (!currentProfile) {
@@ -60,30 +59,30 @@ const PreviewList: FC<PreviewListProps> = ({ className, selectedConversationKey 
     >
       <Card className="mb-6 flex justify-between font-bold">
         <div
-          onClick={() => changeProvider(MESSAGING_PROVIDER.XMTP)}
+          onClick={() => changeChatProvider(MESSAGING_PROVIDER.XMTP)}
           className={`flex basis-1/2 cursor-pointer items-center justify-center rounded-l-xl py-2.5 transition-all hover:bg-gray-200 ${
-            provider === MESSAGING_PROVIDER.XMTP && 'bg-gray-100'
+            chatProvider === MESSAGING_PROVIDER.XMTP && 'bg-gray-100'
           }`}
         >
           <img width={16} height={16} className="mx-1" src="/xmtp.svg" alt="xmtp" draggable={false} />
-          <Trans>{MESSAGING_PROVIDER.XMTP}</Trans>
+          <Trans>{MESSAGING_PROVIDER.XMTP.toUpperCase()}</Trans>
         </div>
         <div
-          onClick={() => changeProvider(MESSAGING_PROVIDER.PUSH)}
+          onClick={() => changeChatProvider(MESSAGING_PROVIDER.PUSH)}
           className={`flex basis-1/2 cursor-pointer items-center justify-center rounded-r-xl py-2.5 transition-all hover:bg-gray-200 ${
-            provider === MESSAGING_PROVIDER.PUSH && 'bg-gray-100'
+            chatProvider === MESSAGING_PROVIDER.PUSH && 'bg-gray-100'
           }`}
         >
           <img width={20} height={20} className="mx-1" src="/push.svg" alt="xmtp" draggable={false} />
-          <Trans>{MESSAGING_PROVIDER.PUSH}</Trans>
+          <Trans>{MESSAGING_PROVIDER.PUSH.toUpperCase()}</Trans>
         </div>
       </Card>
 
       <div className="flex h-[91.8%] flex-col justify-between">
-        {provider === MESSAGING_PROVIDER.XMTP ? (
+        {chatProvider === MESSAGING_PROVIDER.XMTP ? (
           <XMTPPreviewList selectedConversationKey={selectedConversationKey} />
         ) : (
-          <PushPreviewList selectedConversationKey={selectedConversationKey} />
+          <PUSHPreviewList selectedConversationKey={selectedConversationKey} />
         )}
       </div>
     </GridItemFour>
