--- apps/web/src/components/PushMessage/PreviewList.tsx
@@ -1,8 +1,0 @@
-// import React from 'react';
-const PreviewList = () => {
-  // const router = useRouter();
-  // const currentProfile = useAppStore((state) => state.currentProfile);
-};
-export default PreviewList;
-
-// export default PreviewList;
