+++ apps/web/src/pages/messages/xmtp/[...conversationKey].tsx
@@ -1,3 +1,3 @@
-import Message from '@components/Messages/Message';
+import Message from '@components/Messages/Xmtp/Message';
 
 export default Message;
